Fix grammar and making some things more clear in README.md
I'm not sure if I got everything too.

Unclear things not fixed in this PR:
"If you enter 0 and after enter anything other than a dot (for example, 2), it will display a number that you entered (for example, 2)" would be better of written as "If you input 0 and then enter a number, it will replace the zero with a number", because entering math operations does math operations with zero, and math operations are buttons other than dot, making things unclear.
I have no clue "what simply cutted" means? Could be written much better. Perhaps use "displayed for the fixed length of 18 characters as to fit into the calculator boundaries?"